Added CineMachine Camera & deletion of lazer
CineMachine is a way to improve camera following , smoothing and etc. Take a look see at it. Look at the body part of the cinemachine to mess with smoothing and space for the ship to fly in. I added a line of code to delete a lazer when it comes in contact with a asteroid.